# week6/DL-07-Agentic-AI-System-II/02_api_backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminJobPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminJob"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AdminJobPage",
        "type": "object"
      },
...
      "FeedbackOutcome": {
        "enum": [
          "FOLLOWED",
          "IGNORED",
          "CHANGED_PLAN",
          "UNKNOWN"
        ],
        "title": "FeedbackOutcome",
        "type": "string"
      },
      "FieldErrorItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldErrorItem",
        "type": "object"
      },
...
      "JobError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "type",
          "status",
          "stage",
          "progress",
          "created_at",
          "updated_at",
          "result_url",
          "error"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStage": {
        "enum": [
          "queued",
          "fetching_data",
          "assessing_risk",
          "generating_advice",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStage",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobType": {
        "enum": [
          "RECOMMENDATION",
          "MESSAGE",
          "TRIP_ASSESSMENT",
          "DATA_EXPORT"
        ],
        "title": "JobType",
        "type": "string"
      },
...
      "ProblemResponse": {
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details \u2014 the shape of every error response (app/api/problem.py).\n\nFastAPI documents its own `HTTPValidationError` for 422 by default; `custom_openapi`\nin app/main.py replaces every generated 422 (and adds this for other error statuses\ndeclared per-route) with this schema so the exported contract matches what the API\nactually sends.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "request_id",
          "correlation_id",
          "errors"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
      "ReportType": {
        "enum": [
          "UNSAFE_ADVICE",
          "INCORRECT_INFO",
          "OUTDATED_INFO",
          "OTHER"
        ],
        "title": "ReportType",
        "type": "string"
      },
...
      "ReviewDecision": {
        "enum": [
          "approved",
          "rejected"
        ],
        "title": "ReviewDecision",
        "type": "string"
      },
      "ReviewPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ReviewPage",
        "type": "object"
      },
      "ReviewQueueItem": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback_id": {
            "format": "uuid",
            "title": "Feedback Id",
            "type": "string"
          },
          "helpful": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Helpful"
          },
          "outcome": {
            "$ref": "#/components/schemas/FeedbackOutcome"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "recommendation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "recommendation_id": {
            "format": "uuid",
            "title": "Recommendation Id",
            "type": "string"
          },
          "report_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportType"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "required": [
          "feedback_id",
          "recommendation_id",
          "rating",
          "helpful",
          "outcome",
          "report_type",
          "comment",
          "review_status",
          "review_note",
          "reviewed_at",
          "created_at",
          "recommendation"
        ],
        "title": "ReviewQueueItem",
        "type": "object"
      },
      "ReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback_id": {
            "format": "uuid",
            "title": "Feedback Id",
            "type": "string"
          },
          "helpful": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Helpful"
          },
          "outcome": {
            "$ref": "#/components/schemas/FeedbackOutcome"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "recommendation_id": {
            "format": "uuid",
            "title": "Recommendation Id",
            "type": "string"
          },
          "report_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportType"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "required": [
          "feedback_id",
          "recommendation_id",
          "rating",
          "helpful",
          "outcome",
          "report_type",
          "comment",
          "review_status",
          "review_note",
          "reviewed_at",
          "created_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
      "ReviewStatus": {
        "enum": [
          "not_required",
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ReviewStatus",
        "type": "string"
      },
      "ReviewUpdate": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewDecision"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReviewUpdate",
        "type": "object"
      },
...
      "StreamTicketResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "ticket": {
            "title": "Ticket",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "expires_in"
        ],
        "title": "StreamTicketResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Travel Safety & Advisory API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/feedback/reviews": {
      "get": {
        "operationId": "list_reviews_v1_admin_feedback_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReviewStatus",
              "default": "pending"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Feedback waiting for review, oldest first",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/feedback/reviews/{feedback_id}": {
      "patch": {
        "operationId": "review_feedback_v1_admin_feedback_reviews__feedback_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Approve or reject reported feedback",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/jobs": {
      "get": {
        "operationId": "list_jobs_v1_admin_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminJobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Jobs in a time range (default: the last 24 hours)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/jobs/{job_id}": {
      "delete": {
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Cancel a job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Job status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ticket",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticket"
            }
          },
          {
            "in": "header",
            "name": "last-event-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Job progress as Server-Sent Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/stream-ticket": {
      "post": {
        "operationId": "create_stream_ticket_v1_jobs__job_id__stream_ticket_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Short-lived ticket for EventSource clients",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}